Move setenv() and running_as_admin() to platform namespace
