+++ src/util/platform.h
@@ -28,4 +28,15 @@ extern bool isTTY();
  */
 extern std::vector<filesystem::path> glob(const filesystem::path& pattern, size_t max = 100);
 
+/**
+ * Platform specific-implementation of setenv(). Follows the same semantics as
+ * POSIX's setenv() on all platforms.
+ */
+extern int setenv(const char* name, const char* value, int overwrite);
+
+/**
+ * Checks for whether the process is running with administrator rights.
+ */
+extern bool runningAsAdmin();
+
 } // namespace zeek::agent::platform
